+++ src/xmat-models/xmat-http-config.model.ts
@@ -1,8 +1,9 @@
 import {HttpParams} from "@angular/common/http";
 import {XmatDelayedHttpParams} from "./xmat-delayed-http-params.model";
+import {XmatRestVerbs} from "./xmat-rest-verbs.model";
 
 export interface XmatHttpConfig {
-    method: string;
+    method: XmatRestVerbs;
     url: string;
     data?: object;
     queryable: boolean;