+++ src/xmat-models/xmat-rest-verbs.model.ts
@@ -1,6 +1,9 @@
-export interface XmatRestVerbs {
-    GET: string;
-    POST: string;
-    PUT: string;
-    DELETE: string;
+export enum XmatRestVerbs {
+    GET = "GET",
+    POST = "POST",
+    PUT = "PUT",
+    DELETE = "DELETE"
 }
+
+export type XmatDynamicRestVerbsRef<T> = {[key in keyof typeof XmatRestVerbs]: T};
+